+++ realm/realm-annotations-processor/src/main/java/io/realm/processor/Constants.java
@@ -16,10 +16,7 @@
 
 package io.realm.processor;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 public class Constants {