+++ realm/realm-annotations-processor/src/main/java/io/realm/processor/RealmProxyClassGenerator.java
@@ -463,13 +463,14 @@ public class RealmProxyClassGenerator {
                     }
                     writer.endControlFlow();
                 } else {
-                    // check if old Realm has nullable PrimaryKey field and a null value in the field.
+                    // check before migrating a nullable field containing null value to not-nullable PrimaryKey field for Realm version 0.89+
                     if (field.equals(metadata.getPrimaryKey())) {
                         writer
                             .beginControlFlow("if (table.isColumnNullable(%s) && table.findFirstNull(%s) != TableOrView.NO_MATCH)",
                                     fieldIndexVariableReference(field), fieldIndexVariableReference(field))
-                            .emitStatement("throw new IllegalStateException(\"Field '%s' contains null value. Cannot convert objects with a null primary key value.\")",
-                                    fieldName)
+                            .emitStatement("throw new IllegalStateException(\"Cannot migrate an object with null value in field '%s'." +
+                                    " Either maintain the same type for primary key field '%s', or remove the object with null value before migration.\")",
+                                    fieldName, fieldName)
                             .endControlFlow();
                     } else {
                         writer.beginControlFlow("if (table.isColumnNullable(%s))", fieldIndexVariableReference(field));